+++ packages/pluggableWidgets/signature-web/CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+## [2.1.0] - 2026-08-28
+
+### Fixed
+
+- We fixed an issue where strokes stopped being registered in some cases.
+
+- We fixed an issue where the signature canvas was initialized at the wrong size and did not fill its container.
+
 ## [2.0.1] - 2026-07-17
 
 ### Fixed